Refactor: Place 상세 ViewData #184
