+++ lib/features/place/presentation/providers/place_detail_view_provider.dart
@@ -1,17 +1,18 @@
 import 'package:commonplant_frontend/core/config/app_environment.dart';
-import 'package:commonplant_frontend/features/place/domain/entities/place_summary.dart';
 import 'package:commonplant_frontend/features/place/presentation/fixtures/place_detail_fixture.dart';
+import 'package:commonplant_frontend/features/place/presentation/mappers/place_detail_view_mapper.dart';
 import 'package:commonplant_frontend/features/place/presentation/models/place_detail_role.dart';
+import 'package:commonplant_frontend/features/place/presentation/models/place_detail_view_data.dart';
 import 'package:commonplant_frontend/features/place/presentation/providers/place_detail_remote_provider.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 typedef PlaceDetailViewRequest = ({String placeId, PlaceDetailRole? role});
 
 final placeDetailViewProvider =
-    Provider.family<
-      AsyncValue<PlaceDetailFixtureData?>,
-      PlaceDetailViewRequest
-    >((ref, request) {
+    Provider.family<AsyncValue<PlaceDetailViewData?>, PlaceDetailViewRequest>((
+      ref,
+      request,
+    ) {
       if (!ref.watch(useRemoteApiProvider)) {
         return AsyncData(ref.watch(placeLocalDetailViewProvider(request)));
       }
@@ -20,7 +21,7 @@ final placeDetailViewProvider =
     });
 
 final placeLocalDetailViewProvider =
-    Provider.family<PlaceDetailFixtureData, PlaceDetailViewRequest>((
+    Provider.family<PlaceDetailViewData, PlaceDetailViewRequest>((
       ref,
       request,
     ) {
@@ -28,31 +29,20 @@ final placeLocalDetailViewProvider =
     });
 
 final placeRemoteDetailViewProvider =
-    FutureProvider.family<PlaceDetailFixtureData?, PlaceDetailViewRequest>((
+    FutureProvider.family<PlaceDetailViewData?, PlaceDetailViewRequest>((
       ref,
       request,
     ) async {
-      final fixture = ref.watch(placeLocalDetailViewProvider(request));
+      final fallback = ref.watch(placeLocalDetailViewProvider(request));
       final summary = await ref.watch(
         placeDetailProvider(request.placeId).future,
       );
 
-      return applyRemotePlaceSummaryToFixture(
-        fixture: fixture,
+      return mapPlaceSummaryToDetailViewData(
+        fallback: fallback,
         summary: summary,
       );
     }, retry: (retryCount, error) => null);
-
-PlaceDetailFixtureData? applyRemotePlaceSummaryToFixture({
-  required PlaceDetailFixtureData fixture,
-  required PlaceSummary summary,
-}) {
-  if (summary.name.trim().isEmpty) {
-    return null;
-  }
-
-  return fixture.applySummary(summary);
-}
 
 void invalidatePlaceDetailView(WidgetRef ref, PlaceDetailViewRequest request) {
   ref.invalidate(placeDetailProvider(request.placeId));
