+++ test/features/place/presentation/fixtures/place_detail_fixture_test.dart
@@ -1,4 +1,3 @@
-import 'package:commonplant_frontend/features/place/domain/entities/place_summary.dart';
 import 'package:commonplant_frontend/features/place/presentation/fixtures/place_detail_fixture.dart';
 import 'package:commonplant_frontend/features/place/presentation/models/place_detail_role.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -23,23 +22,6 @@ void main() {
       expect(leaderDetail.plants, hasLength(4));
       expect(leaderDetail.plants.first.canWater, isTrue);
       expect(memberDetail.role, PlaceDetailRole.member);
-    });
-
-    test('remote summary를 적용해도 fixture 보조 정보는 유지한다', () {
-      final detail = placeDetailFixture('place-1').applySummary(
-        const PlaceSummary(
-          id: 'remote-place',
-          name: '옥상 정원',
-          address: '서울시 성북구',
-        ),
-      );
-
-      expect(detail.name, '옥상 정원');
-      expect(detail.address, '서울시 성북구');
-      expect(detail.sunlightLabel, '9.3 / 5');
-      expect(detail.humidityLabel, '69%');
-      expect(detail.friends, hasLength(3));
-      expect(detail.plants, hasLength(4));
     });
   });
 }
